// server/vance-addon-brain-binder/client/src/generated/binder/BinderView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.binder.BinderView
*/

import { BinderEntryView } from './BinderEntryView';

export interface BinderView {
  folder: string;
  title?: string;
  description?: string;
  landingRef?: string;
  entries: BinderEntryView[];
}